feat: impl contextPropagationOnly config var; change disableSend behaviour
Implement contextPropagationOnly per spec (it does what disableSend
did before this change). Change disableSend to just disable comms
with APM server, but otherwise *not* attempt to reduce work.
https://github.com/elastic/apm/blob/master/specs/agents/transport.md#context_propagation_only-configuration
