+++ CHANGELOG.asciidoc
@@ -36,6 +36,15 @@ Notes:
 [float]
 ===== Features
 
+* Add <<context-propagation-only, `contextPropagationOnly`>> configuration
+  option. This supports the use case of using the APM agent to propagate HTTP
+  trace-context and to support log-correlation (adding `trace.id` et al fields
+  to log records) **without** an APM server, and to otherwise reduce the
+  processing time of the agent. ({issues}2393[#2393])
++
+This is similar to <<disable-send, `disableSend`>>, but differs in that
+`contextPropagationOnly: true` tells the agent to skip unnecessary work.
+
 [float]
 ===== Bug fixes
 
@@ -256,6 +265,7 @@ The fields affected by `longFieldMaxLength` are:
 
 * Add instrumentation of all AWS S3 methods when using the
   https://www.npmjs.com/package/aws-sdk[JavaScript AWS SDK v2] (`aws-sdk`).
+
 * Add <<disable-send, `disableSend`>> configuration option. This supports some
   use cases using the APM agent **without** an APM server. ({issues}2101[#2101])
