+++ lib/agent.js
@@ -408,8 +408,8 @@ Agent.prototype.captureError = function (err, opts, cb) {
 
   const id = errors.generateErrorId()
 
-  // Quick out if disableSend=true, no point in the processing time.
-  if (this._conf.disableSend) {
+  // Avoid unneeded error/stack processing if only propagating trace-context.
+  if (this._conf.contextPropagationOnly) {
     if (cb) {
       process.nextTick(cb, null, id)
     }
