+++ lib/config.js
@@ -53,6 +53,7 @@ var DEFAULTS = {
   centralConfig: true,
   cloudProvider: 'auto',
   containerId: undefined,
+  contextPropagationOnly: false,
   disableInstrumentations: [],
   disableSend: false,
   environment: process.env.NODE_ENV || 'development',
@@ -109,8 +110,9 @@ var ENV_TABLE = {
   captureHeaders: 'ELASTIC_APM_CAPTURE_HEADERS',
   captureSpanStackTraces: 'ELASTIC_APM_CAPTURE_SPAN_STACK_TRACES',
   centralConfig: 'ELASTIC_APM_CENTRAL_CONFIG',
-  containerId: 'ELASTIC_APM_CONTAINER_ID',
   cloudProvider: 'ELASTIC_APM_CLOUD_PROVIDER',
+  containerId: 'ELASTIC_APM_CONTAINER_ID',
+  contextPropagationOnly: 'ELASTIC_APM_CONTEXT_PROPAGATION_ONLY',
   disableInstrumentations: 'ELASTIC_APM_DISABLE_INSTRUMENTATIONS',
   disableSend: 'ELASTIC_APM_DISABLE_SEND',
   environment: 'ELASTIC_APM_ENVIRONMENT',
@@ -175,6 +177,7 @@ var BOOL_OPTS = [
   'captureHeaders',
   'captureSpanStackTraces',
   'centralConfig',
+  'contextPropagationOnly',
   'disableSend',
   'errorOnAbortedRequests',
   'filterHttpHeaders',
@@ -279,7 +282,7 @@ class Config {
       this.metricsInterval = 0
     }
 
-    if (this.disableSend) {
+    if (this.disableSend || this.contextPropagationOnly) {
       this.transport = function createNoopTransport (conf, agent) {
         return new NoopTransport()
       }
