+++ lib/instrumentation/index.js
@@ -270,11 +270,12 @@ Instrumentation.prototype.addEndedTransaction = function (transaction) {
     this._log.debug({ ctxmgr: this._runCtxMgr.toString() }, 'addEndedTransaction(%s)', transaction.name)
   }
 
-  if (agent._conf.disableSend) {
-    // Save effort if disableSend=true. This one log.trace related to
-    // disableSend is included as a possible log hint to future debugging for
-    // why events are not being sent to APM server.
-    agent.logger.trace('disableSend: skip sendTransaction')
+  // Avoid transaction filtering time if only propagating trace-context.
+  if (agent._conf.contextPropagationOnly) {
+    // This one log.trace related to contextPropagationOnly is included as a
+    // possible log hint to future debugging for why events are not being sent
+    // to APM server.
+    agent.logger.trace('contextPropagationOnly: skip sendTransaction')
     return
   }
 
@@ -306,7 +307,8 @@ Instrumentation.prototype.addEndedSpan = function (span) {
   }
   this._log.debug({ ctxmgr: this._runCtxMgr.toString() }, 'addEndedSpan(%s)', span.name)
 
-  if (agent._conf.disableSend) {
+  // Avoid span encoding time if only propagating trace-context.
+  if (agent._conf.contextPropagationOnly) {
     return
   }
