+++ lib/instrumentation/transaction.js
@@ -288,8 +288,8 @@ Transaction.prototype._captureBreakdown = function (span) {
   const metrics = agent._metrics
   const conf = agent._conf
 
-  // Quick out if disableSend=true, no point in the processing time.
-  if (conf.disableSend) {
+  // Avoid unneeded breakdown metrics processing if only propagating trace context.
+  if (conf.contextPropagationOnly) {
     return
   }
