+++ lib/metrics/index.js
@@ -23,7 +23,7 @@ class Metrics {
 
   start (refTimers) {
     const metricsInterval = this[agentSymbol]._conf.metricsInterval
-    const enabled = metricsInterval !== 0 && !this[agentSymbol]._conf.disableSend
+    const enabled = metricsInterval !== 0 && !this[agentSymbol]._conf.contextPropagationOnly
     if (enabled) {
       this[registrySymbol] = new MetricsRegistry(this[agentSymbol], {
         reporterOptions: {
