+++ lib/noop-transport.js
@@ -1,7 +1,10 @@
 'use strict'
 
 // A no-op (does nothing) Agent transport -- i.e. the APM server client API
-// provided by elastic-apm-http-client. This is used when `disableSend=true`.
+// provided by elastic-apm-http-client.
+//
+// This is used for some configurations (when `disableSend=true` or when
+// `contextPropagationOnly=true`) and in some tests.
 
 class NoopTransport {
   config (opts) {}
